Document ControllableDeployer as only able to return NodeLocalStates.
All current users of this class only use it to pass NodeStates down from
discover_state. If there is a need for a ControllableDeployer that can return
arbitrary ILocalStates from discover_state we can add it later.

Alternatively, we probably don't have to re-invent 1-off mocks and can just use
pymock instead if the need arises.